renaming handshake control as that makes no sense
