+++ src/session/protobuf_construction.h
